cuda working

+++ apps/hello_world/cuda/device_kernels.cuh
@@ -8,9 +8,9 @@ namespace cuda {
 
 namespace kernels {
 
-__global__ void vector_add(const int *input_a,
-                           const int *input_b,
-                           int *output,
+__global__ void vector_add(const float *input_a,
+                           const float *input_b,
+                           float *output,
                            size_t start,
                            size_t end);
